JSON parser/printer completed
